requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Classful==0.14.2
futures==2.2.0
goslate==1.5.2
gTTS==2.2.3
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
lxml==4.6.4
MarkupSafe==2.0.1
Pillow==8.4.0
PyDictionary==2.0.1
pygame==1.9.6
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
urllib3==1.26.7
Werkzeug==2.0.2
